+++ CHANGELOG.md
@@ -7,6 +7,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
+## Version 1.12.2, 6/14/2019
+
+### Added
+
+N/A
+
+### Removed
+
+N/A
+
+### Changed
+
+Disable HTML encape sequence for GSON serializer
+
+---
+
 ## Version 1.12.1, 6/10/2019
 
 ### Added